[FIXED] Returning i64 instead of i32 for large numbers

+++ src/days/day01.rs
@@ -1,13 +1,13 @@
 use crate::utils::VecStringTrim;
 use std::collections::HashMap;
 
-pub fn s1(input: Vec<String>) -> i32 {
+pub fn s1(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
-    let (mut left, mut right): (Vec<i32>, Vec<i32>) = (vec![], vec![]);
+    let (mut left, mut right): (Vec<i64>, Vec<i64>) = (vec![], vec![]);
     for row in input {
         let mut items = row
             .split("   ")
-            .map(|item| item.parse::<i32>().unwrap())
+            .map(|item| item.parse::<i64>().unwrap())
             .into_iter();
         left.push(items.next().unwrap());
         right.push(items.next().unwrap());
@@ -20,15 +20,15 @@ pub fn s1(input: Vec<String>) -> i32 {
         .sum()
 }
 
-pub fn s2(input: Vec<String>) -> i32 {
+pub fn s2(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
-    let mut left_map: HashMap<i32, (i32, i32)> = HashMap::new();
-    let right_items: Vec<i32> = input
+    let mut left_map: HashMap<i64, (i64, i64)> = HashMap::new();
+    let right_items: Vec<i64> = input
         .into_iter()
         .map(|row| {
             let mut pair = row
                 .split("   ")
-                .map(|item| item.parse::<i32>().unwrap())
+                .map(|item| item.parse::<i64>().unwrap())
                 .into_iter();
             let first = pair.next().unwrap();
             if let Some((left_times, _)) = left_map.get_mut(&first) {
