+++ src/days/day05.rs
@@ -2,8 +2,8 @@ use crate::utils::VecStringTrim;
 use std::collections::HashMap;
 
 struct PagesHelper {
-    pages_at_left: Vec<i32>,
-    pages_at_right: Vec<i32>,
+    pages_at_left: Vec<i64>,
+    pages_at_right: Vec<i64>,
 }
 
 impl PagesHelper {
@@ -14,19 +14,19 @@ impl PagesHelper {
         }
     }
 
-    fn add_to_left(&mut self, value: i32) {
+    fn add_to_left(&mut self, value: i64) {
         self.pages_at_left.push(value);
     }
 
-    fn add_to_right(&mut self, value: i32) {
+    fn add_to_right(&mut self, value: i64) {
         self.pages_at_right.push(value);
     }
 
-    fn get_pages_at_left(&self) -> Vec<i32> {
+    fn get_pages_at_left(&self) -> Vec<i64> {
         self.pages_at_left.clone()
     }
 
-    fn get_pages_at_right(&self) -> Vec<i32> {
+    fn get_pages_at_right(&self) -> Vec<i64> {
         self.pages_at_right.clone()
     }
 }
@@ -51,17 +51,17 @@ fn get_page_number_updates(input: Vec<String>) -> Vec<String> {
     input.into_iter().skip(hash_at + 1).collect()
 }
 
-pub fn s1(input: Vec<String>) -> i32 {
+pub fn s1(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
     let page_order_rules = get_page_ordering_rules(input.clone());
     let page_numbers = get_page_number_updates(input);
 
-    let mut page_order_map: HashMap<i32, PagesHelper> = HashMap::new();
+    let mut page_order_map: HashMap<i64, PagesHelper> = HashMap::new();
 
     for pg in page_order_rules {
         let mut pgs = pg.split("|");
-        let at_left: i32 = pgs.next().unwrap().parse().unwrap();
-        let at_right: i32 = pgs.next().unwrap().parse().unwrap();
+        let at_left: i64 = pgs.next().unwrap().parse().unwrap();
+        let at_right: i64 = pgs.next().unwrap().parse().unwrap();
         if let Some(page) = page_order_map.get_mut(&at_left) {
             page.add_to_right(at_right);
         } else {
@@ -82,7 +82,7 @@ pub fn s1(input: Vec<String>) -> i32 {
     let correct_pages: Vec<String> = page_numbers
         .into_iter()
         .filter(|row| {
-            let pgs: Vec<i32> = row.split(",").map(|v| v.parse().unwrap()).collect();
+            let pgs: Vec<i64> = row.split(",").map(|v| v.parse().unwrap()).collect();
             for i in 0..pgs.len() {
                 let cur_page = pgs[i];
                 let previous = &pgs[..i];
@@ -104,10 +104,10 @@ pub fn s1(input: Vec<String>) -> i32 {
         })
         .collect();
 
-    let middles: Vec<i32> = correct_pages
+    let middles: Vec<i64> = correct_pages
         .into_iter()
         .map(|row| {
-            let parsed: Vec<i32> = row.split(",").map(|v| v.parse().unwrap()).collect();
+            let parsed: Vec<i64> = row.split(",").map(|v| v.parse().unwrap()).collect();
             let targ = (parsed.len() as f64 / 2.0).round() as usize - 1;
             parsed[targ]
         })
@@ -116,17 +116,17 @@ pub fn s1(input: Vec<String>) -> i32 {
     middles.iter().sum()
 }
 
-pub fn s2(input: Vec<String>) -> i32 {
+pub fn s2(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
     let page_order_rules = get_page_ordering_rules(input.clone());
     let page_numbers = get_page_number_updates(input);
 
-    let mut page_order_map: HashMap<i32, PagesHelper> = HashMap::new();
+    let mut page_order_map: HashMap<i64, PagesHelper> = HashMap::new();
 
     for pg in page_order_rules {
         let mut pgs = pg.split("|");
-        let at_left: i32 = pgs.next().unwrap().parse().unwrap();
-        let at_right: i32 = pgs.next().unwrap().parse().unwrap();
+        let at_left: i64 = pgs.next().unwrap().parse().unwrap();
+        let at_right: i64 = pgs.next().unwrap().parse().unwrap();
         if let Some(page) = page_order_map.get_mut(&at_left) {
             page.add_to_right(at_right);
         } else {
@@ -147,7 +147,7 @@ pub fn s2(input: Vec<String>) -> i32 {
     let incorrect_pages: Vec<String> = page_numbers
         .into_iter()
         .filter(|row| {
-            let pgs: Vec<i32> = row.split(",").map(|v| v.parse().unwrap()).collect();
+            let pgs: Vec<i64> = row.split(",").map(|v| v.parse().unwrap()).collect();
             for i in 0..pgs.len() {
                 let cur_page = pgs[i];
                 let previous = &pgs[..i];
@@ -169,15 +169,15 @@ pub fn s2(input: Vec<String>) -> i32 {
         })
         .collect();
 
-    let incorrects_as_vec: Vec<Vec<i32>> = incorrect_pages
+    let incorrects_as_vec: Vec<Vec<i64>> = incorrect_pages
         .into_iter()
         .map(|row| row.split(",").map(|v| v.parse().unwrap()).collect())
         .collect();
 
-    let mut correcting: Vec<Vec<i32>> = Vec::new();
+    let mut correcting: Vec<Vec<i64>> = Vec::new();
 
     for vals in incorrects_as_vec.into_iter() {
-        let mut this_row: Vec<i32> = Vec::new();
+        let mut this_row: Vec<i64> = Vec::new();
 
         for v in vals {
             let target_ind = this_row.iter().enumerate().fold(0, |ini, (i, elem)| {
